implements levels on units and made damage and health doubles

+++ src/main/java/test/game/model/entity/EntityTest.java
@@ -47,7 +47,7 @@ public class EntityTest {
 
     World world = getWorld();
     player.setTargetUnit(enemy);
-    int prevHealth = enemy.getHealth();
+    double prevHealth = enemy.getHealth();
     for (int i = 0; i < 900; i++) {
       player.tick(GameModel.DELAY, world);
     }
